+++ command/cp.go
@@ -8,6 +8,7 @@ import (
 	"net/http"
 	"os"
 	"strings"
+	"time"
 
 	"github.com/hashicorp/go-multierror"
 	"github.com/urfave/cli/v2"
@@ -15,6 +16,7 @@ import (
 	errorpkg "github.com/peak/s5cmd/error"
 	"github.com/peak/s5cmd/log"
 	"github.com/peak/s5cmd/parallel"
+	"github.com/peak/s5cmd/statutil"
 	"github.com/peak/s5cmd/storage"
 	"github.com/peak/s5cmd/storage/url"
 )
@@ -191,7 +193,8 @@ increase the open file limit or try to decrease the number of workers with
 `
 
 // Run starts copying given source objects to destination.
-func (c Copy) Run(ctx context.Context) error {
+func (c Copy) Run(ctx context.Context) (err error) {
+	defer statutil.StatCollect("Copy.Run", time.Now(), &err)()
 	srcurl, err := url.New(c.src)
 	if err != nil {
 		return err
@@ -341,11 +344,12 @@ func (c Copy) prepareUploadTask(
 }
 
 // doDownload is used to fetch a remote object and save as a local object.
-func (c Copy) doDownload(ctx context.Context, srcurl *url.URL, dsturl *url.URL) error {
+func (c Copy) doDownload(ctx context.Context, srcurl *url.URL, dsturl *url.URL) (err error) {
+	defer statutil.StatCollect("Copy.doDownload", time.Now(), &err)()
 	srcClient := storage.NewClient(srcurl)
 	dstClient := storage.NewClient(dsturl)
 
-	err := c.shouldOverride(ctx, srcurl, dsturl)
+	err = c.shouldOverride(ctx, srcurl, dsturl)
 	if err != nil {
 		// FIXME(ig): rename
 		if errorpkg.IsWarning(err) {
@@ -384,7 +388,8 @@ func (c Copy) doDownload(ctx context.Context, srcurl *url.URL, dsturl *url.URL) 
 	return nil
 }
 
-func (c Copy) doUpload(ctx context.Context, srcurl *url.URL, dsturl *url.URL) error {
+func (c Copy) doUpload(ctx context.Context, srcurl *url.URL, dsturl *url.URL) (err error) {
+	defer statutil.StatCollect("Copy.doUpload", time.Now(), &err)()
 	// TODO(ig): use storage abstraction
 	f, err := os.Open(srcurl.Absolute())
 	if err != nil {
@@ -442,7 +447,8 @@ func (c Copy) doUpload(ctx context.Context, srcurl *url.URL, dsturl *url.URL) er
 	return nil
 }
 
-func (c Copy) doCopy(ctx context.Context, srcurl *url.URL, dsturl *url.URL) error {
+func (c Copy) doCopy(ctx context.Context, srcurl *url.URL, dsturl *url.URL) (err error) {
+	defer statutil.StatCollect("Copy.doCopy", time.Now(), &err)()
 	srcClient := storage.NewClient(srcurl)
 
 	metadata := storage.NewMetadata().
@@ -451,7 +457,7 @@ func (c Copy) doCopy(ctx context.Context, srcurl *url.URL, dsturl *url.URL) erro
 		SetSSEKeyID(c.encryptionKeyID).
 		SetACL(c.acl)
 
-	err := c.shouldOverride(ctx, srcurl, dsturl)
+	err = c.shouldOverride(ctx, srcurl, dsturl)
 	if err != nil {
 		if errorpkg.IsWarning(err) {
 			printDebug(c.op, srcurl, dsturl, err)