+++ command/ls.go
@@ -3,12 +3,14 @@ package command
 import (
 	"context"
 	"fmt"
+	"time"
 
 	"github.com/hashicorp/go-multierror"
 	"github.com/urfave/cli/v2"
 
 	errorpkg "github.com/peak/s5cmd/error"
 	"github.com/peak/s5cmd/log"
+	"github.com/peak/s5cmd/statutil"
 	"github.com/peak/s5cmd/storage"
 	"github.com/peak/s5cmd/storage/url"
 	"github.com/peak/s5cmd/strutil"
@@ -110,7 +112,8 @@ func List(
 	showEtag bool,
 	humanize bool,
 	showStorageClass bool,
-) error {
+) (err error) {
+	defer statutil.StatCollect("List", time.Now(), &err)()
 	srcurl, err := url.New(src)
 	if err != nil {
 		return err