+++ storage/s3.go
@@ -25,6 +25,7 @@ import (
 	"github.com/aws/aws-sdk-go/service/s3/s3manager"
 	"github.com/aws/aws-sdk-go/service/s3/s3manager/s3manageriface"
 
+	"github.com/peak/s5cmd/statutil"
 	"github.com/peak/s5cmd/storage/url"
 )
 
@@ -109,7 +110,8 @@ func NewS3Storage(opts S3Options) (*S3, error) {
 }
 
 // Stat retrieves metadata from S3 object without returning the object itself.
-func (s *S3) Stat(ctx context.Context, url *url.URL) (*Object, error) {
+func (s *S3) Stat(ctx context.Context, url *url.URL) (obj *Object, err error) {
+	defer statutil.StatCollect("S3.Stat", time.Now(), &err)()
 	output, err := s.api.HeadObjectWithContext(ctx, &s3.HeadObjectInput{
 		Bucket: aws.String(url.Bucket),
 		Key:    aws.String(url.Path),
@@ -317,7 +319,8 @@ func (s *S3) listObjects(ctx context.Context, url *url.URL) <-chan *Object {
 
 // Copy is a single-object copy operation which copies objects to S3
 // destination from another S3 source.
-func (s *S3) Copy(ctx context.Context, from, to *url.URL, metadata Metadata) error {
+func (s *S3) Copy(ctx context.Context, from, to *url.URL, metadata Metadata) (err error) {
+	defer statutil.StatCollect("S3.Copy", time.Now(), &err)()
 	// SDK expects CopySource like "bucket[/key]"
 	copySource := strings.TrimPrefix(from.String(), "s3://")
 
@@ -346,7 +349,7 @@ func (s *S3) Copy(ctx context.Context, from, to *url.URL, metadata Metadata) err
 		input.ACL = aws.String(acl)
 	}
 
-	_, err := s.api.CopyObject(input)
+	_, err = s.api.CopyObject(input)
 	return err
 }
 
@@ -359,7 +362,8 @@ func (s *S3) Get(
 	to io.WriterAt,
 	concurrency int,
 	partSize int64,
-) (int64, error) {
+) (res int64, err error) {
+	defer statutil.StatCollect("S3.Get", time.Now(), &err)()
 	if concurrency == 1 {
 		resp, err := s.api.GetObjectWithContext(ctx, &s3.GetObjectInput{
 			Bucket: aws.String(from.Bucket),
@@ -391,7 +395,8 @@ func (s *S3) Put(
 	metadata Metadata,
 	concurrency int,
 	partSize int64,
-) error {
+) (err error) {
+	defer statutil.StatCollect("S3.Put", time.Now(), &err)()
 	contentType := metadata.ContentType()
 	if contentType == "" {
 		contentType = "application/octet-stream"
@@ -422,7 +427,7 @@ func (s *S3) Put(
 		}
 	}
 
-	_, err := s.uploader.UploadWithContext(ctx, input, func(u *s3manager.Uploader) {
+	_, err = s.uploader.UploadWithContext(ctx, input, func(u *s3manager.Uploader) {
 		u.PartSize = partSize
 		u.Concurrency = concurrency
 	})
@@ -478,7 +483,8 @@ func (s *S3) calculateChunks(ch <-chan *url.URL) <-chan chunk {
 }
 
 // Delete is a single object delete operation.
-func (s *S3) Delete(ctx context.Context, url *url.URL) error {
+func (s *S3) Delete(ctx context.Context, url *url.URL) (err error) {
+	defer statutil.StatCollect("S3.Delete", time.Now(), &err)()
 	chunk := chunk{
 		Bucket: url.Bucket,
 		Keys: []*s3.ObjectIdentifier{
@@ -580,8 +586,9 @@ func (s *S3) ListBuckets(ctx context.Context, prefix string) ([]Bucket, error) {
 }
 
 // MakeBucket creates an S3 bucket with the given name.
-func (s *S3) MakeBucket(ctx context.Context, name string) error {
-	_, err := s.api.CreateBucketWithContext(ctx, &s3.CreateBucketInput{
+func (s *S3) MakeBucket(ctx context.Context, name string) (err error) {
+	defer statutil.StatCollect("S3.MakeBucket", time.Now(), &err)()
+	_, err = s.api.CreateBucketWithContext(ctx, &s3.CreateBucketInput{
 		Bucket: aws.String(name),
 	})
 	return err
@@ -589,7 +596,8 @@ func (s *S3) MakeBucket(ctx context.Context, name string) error {
 
 // NewAwsSession initializes a new AWS session with region fallback and custom
 // options.
-func newSession(opts S3Options) (*session.Session, error) {
+func newSession(opts S3Options) (sess *session.Session, err error) {
+	defer statutil.StatCollect("newSession", time.Now(), &err)()
 	awsCfg := aws.NewConfig()
 
 	endpointURL, err := parseEndpoint(opts.Endpoint)
@@ -638,7 +646,7 @@ func newSession(opts S3Options) (*session.Session, error) {
 		}
 	}
 
-	sess, err := session.NewSessionWithOptions(
+	sess, err = session.NewSessionWithOptions(
 		session.Options{
 			Config:            *awsCfg,
 			SharedConfigState: useSharedConfig,